refactor(provideautospy): change `createSpyFromClass` to the shorter `provideAutoSpy`
Since `jasmine-auto-spies` v6.9.0 and `jest-auto-spies` v1.6.0 there's an ability to use the shorter
util function `provideAutoSpy()`. So upgrade to those versions in order to not experience any import
issues.

BREAKING CHANGE: Since `jasmine-auto-spies` v6.9.0 and `jest-auto-spies` v1.6.0 there's an ability
to use the shorter util function `provideAutoSpy()`. So upgrade to those versions in order to not
experience any import issues.